feat(domain): hardened entity model, rich domain behavior, BigDecimal financials, and constructor injection (Phase 2A) (#4)
- Replaced Double with BigDecimal (precision = 19, scale = 4) for monetary fields in User and Transaction entities
- Implemented rich domain methods User.debit() and User.credit() encapsulating balance invariants
- Added audit timestamps (@CreationTimestamp, @UpdateTimestamp) and optimistic locking (@Version) to User
- Added TransactionStatus and TransactionType enums
- Added JPA @ManyToOne relationships between Transaction and User with denormalized UPI strings
- Added UUID referenceId auto-generation on Transaction (@PrePersist)
- Refactored UserService, TransactionService, UserController, and TransactionController to constructor injection
- Added unit test suite for User domain logic and Transaction reference ID (UserTest, TransactionTest)
- Added ADR-001 (BigDecimal), ADR-002 (Constructor injection), ADR-003 (Rich Domain Model)
- Updated ARCHITECTURE.md, ROADMAP.md, CHANGELOG.md, and README.md

+++ src/main/java/com/payflow/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package com.payflow.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,5 +17,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
 	// JPQL query to find users with balance greater than a specified amount
 	@Query("SELECT u FROM User u WHERE u.balance > :amount")
-	List<User> findUsersWithBalanceGreaterThan(@Param("amount") Double amount);
+	List<User> findUsersWithBalanceGreaterThan(@Param("amount") BigDecimal amount);
 }
